LayoutTests/http/tests/websocket/tests/hybi/handshake-fail-by-maxlength-expected.txt

CONSOLE MESSAGE: WebSocket connection to 'ws://localhost:8880/websocket/tests/hybi/handshake-fail-by-maxlength' failed: Status line is too long
Connection should fail immediately, rather than succeeding or staying in limbo until timeout, if handshake is longer than 1024 bytes.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS timedOut is false
PASS connected is false
PASS origin is undefined.
PASS successfullyParsed is true

TEST COMPLETE
